add flow utils that holds shared logic
